application class docs updated

+++ Cache/Handler/File.php
@@ -207,7 +207,10 @@ class File implements CacheHandlerInterface
      */
     public function delete($key)
     {
-        return unlink($this->filePath . $key);
+        if (file_exists($this->filePath . $key)) {
+            return unlink($this->filePath . $key);
+        }
+        return false;
     }
 
     /**
